start replacing list view
