+++ assets/src/components/cd/cluster/addon/ClusterAddOnCompatibility.tsx
@@ -8,7 +8,7 @@ import { AddonVersion } from 'generated/graphql'
 import { TabularNumbers } from 'components/cluster/TableElements'
 import { FullHeightTableWrap } from 'components/utils/layout/FullHeightTableWrap'
 
-import { useClusterAddOnContext } from './ClusterAddOn'
+import { useClusterAddOnContext } from '../ClusterAddOns'
 
 const Compatibility = memo(
   ({
